11/13 유저 예약목록(확정된 예약) 구현

+++ app/src/main/java/com/blooburn/owere/user/item/UserEntity.kt
@@ -3,9 +3,9 @@ package com.blooburn.owere.user.item
 data class UserEntity(
     val myName : String,
     val userId : String,
-    val latitude: Float,
-    val longitude : Float
+    val latitude: Double,
+    val longitude : Double
 ){  //인자가 없는 경우 추가 필요 (에러)
-    constructor() : this("","",0.0f,0.0f)
+    constructor() : this("","",0.0,0.0)
 }
